+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-Updated to react-places-autocomplete v7.0.1
+* Updated SignInLogic to redux state, but removed the localstorage for now. Upon reload, user gets signed out. ([](https://github.com/CoVoCre/CoVoBETA/commit/))
 
-* Updated LoginAndOtherDropdown to redux state ([66d9b54f](https://github.com/CoVoCre/CoVoBETA/commit/66d9b54f7f0a2c8ed2438e849e94a618864520d3))
+* Updated LoginAndOtherDropdown to redux state ([7ea6f5dd](https://github.com/CoVoCre/CoVoBETA/commit/7ea6f5dd3a4f97681fba80181ab4c9fed89a79f9))
 
 * Added redux-devtools-extension to inspect redux state in chrome with devtolls redux extension ([26931592](https://github.com/CoVoCre/CoVoBETA/commit/269315926b50fdb199967e17aa3292e051a81444))
+
+Updated to react-places-autocomplete v7.0.1